version 0.1.19
New border rendering mechanism! smoother corners and subtle drop shadows will help make the imojis look better than before.
Please note that we've now shifted to using OpenGL to render the borders, which will appear slower than the previous mechanism on the simulator. On the device, performance and memory usage is inline with the old method.